corrected with negatives and without weights
weights cannot be saved like they were: they need to be saved as tf.variables in order to be correctly used for gradients computation. Therefore, there is no need to introduce they weights parameters for loss functions but one can simply save the training weights and then re-instantiate the model and load the saved weights.